workers/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# AI and ML
crewai==0.11.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
openai==1.6.1
anthropic==0.8.1
cohere==4.37

# Vector embeddings
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Messaging and queues
nats-py==2.6.0
redis==5.0.1

# AWS/S3
boto3==1.34.0
minio==7.2.0

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==1.21.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
